requirements.txt
Flask==0.10.1
Flask-Testing==0.4.1
Jinja2==2.7.3
MarkupSafe==0.23
Pillow==2.3.0
Werkzeug==0.9.6
bcrypt==1.0.2
cffi==0.8.2
gunicorn==19.0.0
itsdangerous==0.24
mongoengine==0.8.7
numpy==1.9.0
py-bcrypt==0.4
pycparser==2.10
pycurl==7.19.0
pymongo==2.7.1
python-dateutil==2.2
redis==2.10.1
requests==1.1.0
sendgrid==1.1.2
six==1.7.2
smtpapi==0.1.2
stripe==1.19.0
unittest2==0.5.1
wsgiref==0.1.2
